pyproject.toml
[build-system]
requires = ["setuptools>=77.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "corpus2dnallm"
version = "0.1.0"
description = "A tool for converting genome sequences to language model corpora"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "Corpus2DNALLM Team"}
]
keywords = ["bioinformatics", "genomics", "nlp", "language-model"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.20.0",
    "biopython>=1.79",
    "pyfastx>=0.8.0",
    "sentencepiece>=0.1.96",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.0",
    "flake8>=3.9",
]

[project.scripts]
corpus2dnallm = "corpus2dnallm.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/Corpus2DNALLM"
Repository = "https://github.com/yourusername/Corpus2DNALLM"
Issues = "https://github.com/yourusername/Corpus2DNALLM/issues"


[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
